MOS-15212 - Updated response duration logging
Former-commit-id: ddf22d81accfa957deedc4bf34755baf5a8d6d62
Former-commit-id: fd18acd21e2ba1ec3e0e78b5e7a477f7a1980b7a

+++ authentication/authentication-service/src/main/java/io/mosip/authentication/service/filter/BaseAuthFilter.java
@@ -328,7 +328,7 @@ public abstract class BaseAuthFilter implements Filter {
 				.toMillis();
 		mosipLogger.info(SESSION_ID, EVENT_FILTER, BASE_AUTH_FILTER,
 				"Time difference between request and response in millis:" + duration
-						+ ".  Time difference between request and response in Seconds: " + (duration / 1000));
+						+ ".  Time difference between request and response in Seconds: " + ((double) duration / 1000));
 	}
 
 	/**
